Fix VG article paragraphs.
Redefine the preprocess_raw_html method, and do the substitution in
there instead of using preprocess_regexps.

What the current version does is to replace "<br/><br/>" sequences
with "</p><p>" and then it's left to tagsoup to add the missing
paragraph start and end tags.  Seems to be enough.  But more
substitutions may be added for completeness if needed.