Correctly display images in many nrk.no articles.
In these articles a <figure> element contains both an <img> that uses
JavaScript to display correctly (and the epub viewers won't handle
that), and an <img> wrapped in a <noscript> element that contains a
plain image reference.

Unfortunately the BasicNewsRecipe._postprocess_html strips <noscript>
and all of its children, so the workaround is to strip <noscript> and
</noscript> using a regular expression, and then suppress the
JavaScript <img> in remove_tags (this is the <img class="js-img">
already in remove_tags).