Style the caption of nrk.no images.
This is the caption of the images that were restored by the previous
commit.  Copied the ".image figcaption" and ".image figcaption p"
styles into the new styles ".image div" and ".image div p".  The
matching elements were found by looking at the parsed HTML of the
debug directory (debug/parsed when running ebook-convert with the
"--debug-pipeline debug" arguments).