Reduced the size of the nrk.no epub.
The problem was the size of the "Forsida" feed.

Having max 100 articles per feed gave a size of 16-18MB per epub.
Reducing to 50 articles per feed reduced the epub size to 7MB.
Reducing to 15 articles per feed reduced the epub size to 5.4MB.
Reducing to 10 articles per feed reduced the epub size to 4.6MB (not a big difference).